Improved client error handling
- Fixed scope mismatch when applying moves from the server
- Human always plays next after server (with correct color)

It was just assumed that the server would respond nicely b/c the client didn't know what to do when it didn't.
Currently, all moves from the server will get applied (and not rejected by the client) even if there is a server error. However, the behaviour has been changed so that server errors are just ignored (instead of causing more bugs).
While it might seem that any move which causes a server error should be ignored, this is not the case. Perfectly valid moves by the human can cause an 'object is not iterable' error because the server cannot find a valid move to play (which should be interpereted as a pass).
Attempting to make moves when offline will still cause bugs in the client. However, this should be easily solved by using onOffline an onOnline event handlers to disable and enable the ui.